рефакторинг — разнесение кода по классам
